dinner-party/reply/requirements.txt
# Function dependencies, for example:
# package>=version
twilio
git+https://github.com/anjmittu/dinner-party-database
google-cloud-pubsub
